// lib/gen/api/discourseAPI/schemas/getTopic200.ts
export type GetTopic200 = {
  post_stream: GetTopic200PostStream;
  timeline_lookup: unknown[];
  suggested_topics: GetTopic200SuggestedTopicsItem[];
  tags: unknown[];
  tags_descriptions: GetTopic200TagsDescriptions;
  id: number;
  title: string;
  fancy_title: string;
  posts_count: number;
  created_at: string;
  views: number;
  reply_count: number;
  like_count: number;
  last_posted_at: GetTopic200LastPostedAt;
  visible: boolean;
  closed: boolean;
  archived: boolean;
  has_summary: boolean;
  archetype: string;
  slug: string;
  category_id: number;
  word_count: GetTopic200WordCount;
  deleted_at: GetTopic200DeletedAt;
  user_id: number;
  featured_link: GetTopic200FeaturedLink;
  pinned_globally: boolean;
  pinned_at: GetTopic200PinnedAt;
  pinned_until: GetTopic200PinnedUntil;
  image_url: GetTopic200ImageUrl;
  slow_mode_seconds: number;
  draft: GetTopic200Draft;
  draft_key: string;
  draft_sequence: number;
  unpinned: GetTopic200Unpinned;
  pinned: boolean;
  current_post_number?: number;
  highest_post_number: GetTopic200HighestPostNumber;
  deleted_by: GetTopic200DeletedBy;
  has_deleted: boolean;
  actions_summary: GetTopic200ActionsSummaryItem[];
  chunk_size: number;
  bookmarked: boolean;
  bookmarks: unknown[];
  topic_timer: GetTopic200TopicTimer;
  message_bus_last_id: number;
  participant_count: number;
  show_read_indicator: boolean;
  thumbnails: GetTopic200Thumbnails;
  slow_mode_enabled_until: GetTopic200SlowModeEnabledUntil;
  details: GetTopic200Details;
};